move some files, chart cursor point and background
